requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
backcall==0.2.0
beautifulsoup4==4.7.1
billiard==3.3.0.23
bleach==3.1.5
celery==3.1.26.post2
certifi==2018.11.29
cffi==1.14.2
chardet==3.0.4
Click==7.0
cloudinary==1.13.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
cssutils==1.0.2
decorator==4.4.2
defusedxml==0.5.0
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.4
django-celery==3.3.1
django-cors-headers==2.5.3
django-cors-middleware==1.3.1
django-extensions==2.1.2
django-filter==2.0.0
django-geoposition==0.3.0
django-heroku==0.3.1
django-inlinecss==0.1.2
django-openid-auth==0.15
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
Faker==0.9.1
filelock==3.0.12
flake8==3.5.0
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.19
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.17.1
Jinja2==2.10.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
kombu==3.0.37
lazy-object-proxy==1.4.3
locustio==0.11.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
more-itertools==7.2.0
msgpack==0.6.1
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
openapi-codec==1.3.2
packaging==19.1
pandocfilters==1.4.2
parso==0.7.0
pbr==5.1.3
pickleshare==0.7.5
Pillow==5.3.0
pluggy==0.12.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.8.0
pycodestyle==2.3.1
pycparser==2.20
pyflakes==1.6.0
Pygments==2.6.1
PyJWT==1.6.4
pylint==2.4.4
pynliner3==0.6
pyparsing==2.4.2
pyrsistent==0.16.0
pytest==5.0.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
pywinpty==0.5.7
PyYAML==5.1
pyzmq==18.1.0
qtconsole==4.7.5
QtPy==1.9.0
redis==2.10.6
requests==2.21.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.11.0
soupsieve==1.8
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==20.0.20
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
whitenoise==4.1
widgetsnbextension==3.5.1
wrapt==1.11.2
yapf==0.26.0
zipp==0.5.2
